requirements.txt
python-dotenv==1.0.0
torch==2.0.1
torchvision==0.15.2
diffusers==0.19.3
transformers==4.31.0
accelerate==0.21.0
datasets==2.14.0
pillow==9.5.0
huggingface-hub==0.14.1
numpy<2.0 